Create Mongodb_python_CRUD
This code aims at using the python language for interacting with the MongoDB database. CRUD operations using python language are demonstrated, which manipulates the data stored in MongoDB. PyMongo is a Python distribution containing tools for working with MongoDB. Several applications which may be developed in python may use MongoDB for data storage.